requirements.txt
amqp==2.1.4
appdirs==1.4.3
asn1crypto==0.22.0
bcrypt==3.1.2
billiard==3.5.0.2
celery==4.0.2
certifi==2017.7.27.1
cffi==1.10.0
cryptography==1.8.1
Django==1.8.3
django-annoying==0.10.3
django-localflavor==1.5.2
django-scheduler==0.8.6.1
enum34==1.1.6
httplib2==0.10.3
icalendar==4.0.0
idna==2.5
ipaddress==1.0.18
kombu==4.0.2
oauth==1.0.1
oauth2==1.5.211
oauth2client==1.4.12
oauthlib==2.0.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.2.12
pytz==2016.10
requests==2.11.1
requests-oauth2==0.2.0
requests-oauthlib==0.5.0
rsa==3.4.2
six==1.10.0
twilio==6.0.0rc13
vine==1.1.3
wincertstore==0.2
